:recycle: [BugLifecycle] Changed workflow state name
